pyproject.toml
[tool.poetry]
name = "apples-detection"
version = "0.1.0"
description = "Apple detection and instance segmentation pipeline"
authors = ["Dmitry Nesterov <dim.nesterov2015@gmail.com>"]
readme = "README.md"
packages = [{include = "apples_detection"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<3.11"

# pytorch
pytorch-lightning = "^1.9.3"
torchvision = "^0.14.1"
torchtext = "^0.14.1"
torchmetrics = "^0.11.1"
tensorboard = "^2.12.0"

# hydra
hydra-core = "^1.3.2"
hydra-colorlog = "^1.2.0"
hydra-optuna-sweeper = "^1.2.0"

# other
pyrootutils = "^1.0.4"
rich = "^13.3.1"
wandb = "^0.13.10"
albumentations = "^1.3.0"
opencv-contrib-python = "^4.7.0.72"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
ipython = "^8.10.0"
jupyter = "^1.0.0"
pre-commit = "^3.1.0"
sh = "^2.0.2"
pytest-cov = "^4.0.0"
ipykernel = "^6.21.2"
matplotlib = "^3.7.0"
ruff = "^0.0.254"
black = "^23.1.0"

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.black]
line-length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
